# server/docs/swagger.yaml
...
openapi: 3.0.3
info:
  title: Agro Platform API
  version: 1.0.0
  description: |
    Внутренний и публичный API агро-платформы.
    Основан на PostgreSQL (users, suppliers, places, products и др.).
servers:
  - url: http://localhost:5000
    description: Local development
...
paths:
...
  /api/admin/suppliers:
    get:
      tags: [Admin]
      summary: Получить список поставщиков (админ)
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Список поставщиков с пагинацией
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  suppliers:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminSupplier'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags: [Admin]
      summary: Создать поставщика (админ)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierRegistration'
      responses:
        '201':
          description: Поставщик создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Неавторизованный доступ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
...
    SupplierRegistration:
      type: object
      required: [name, username, email, password]
      properties:
        name:
          type: string
        description:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
...
    AdminSupplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        userid:
          type: integer
        username:
          type: string
        email:
          type: string
        rating:
          type: number
        description:
          type: string
        places_count:
          type: integer